A number of changes related to linear algebra
1. d interface to rawLinAlgSolve changed
2. rawLinAlgSolveInvertible interface added, but not functional yet
3. rawSolve was removed (it was essentially same as rawLinAlgSolve)
4. tests for rank profile, solving were added to EngineTests
5. flint patched so g++ would not complain so much (warning messages)
6. small changes in gf code to remove warnings from g++
7. rawLinAlgRREF commands added for finite fields.
8. dead code removed.

A number of mutable matrix linear algebra functions still don't work
with QQ, and in fact there is an example that fails in this commit.

+++ M2/Macaulay2/e/aring-gf-flint.cpp
@@ -159,7 +159,7 @@ namespace M2 {
   void ARingGFFlint::eval(const RingMap *map, const ElementType& f, int first_var, ring_elem &result) const
   {
     // f is represented by: f.value, the power of the generator
-    result = map->get_ring()->power(map->elem(first_var), f.value);    
+    result = map->get_ring()->power(map->elem(first_var), static_cast<int>(f.value));    
   }
   
   void ARingGFFlint::text_out(buffer &o) const
